Revert "Decouple checkout success page (#811)"
This reverts commit b722ce1f7fa7d92a875e0c2cff5c54e5e7d617e6.
