+++ blocks/commerce-checkout/commerce-checkout.js
@@ -2,6 +2,7 @@
 
 // Dropin Tools
 import { events } from '@dropins/tools/event-bus.js';
+import { initializers } from '@dropins/tools/initializer.js';
 import { initReCaptcha } from '@dropins/tools/recaptcha.js';
 
 // Order Dropin Modules
@@ -31,6 +32,8 @@ import {
 // Fragment functions
 import {
   createCheckoutFragment,
+  createOrderConfirmationFooter,
+  createOrderConfirmationFragment,
   selectors,
 } from './fragments.js';
 
@@ -42,11 +45,18 @@ import {
   renderCartSummaryList,
   renderCheckoutHeader,
   renderCustomerBillingAddresses,
+  renderCustomerDetails,
   renderCustomerShippingAddresses,
   renderEmptyCart,
   renderGiftOptions,
   renderLoginForm,
   renderMergedCartBanner,
+  renderOrderConfirmationFooterButton,
+  renderOrderCostSummary,
+  renderOrderGiftOptions,
+  renderOrderHeader,
+  renderOrderProductList,
+  renderOrderStatus,
   renderOrderSummary,
   renderOutOfStock,
   renderPaymentMethods,
@@ -54,6 +64,7 @@ import {
   renderServerError,
   renderShippingAddressFormSkeleton,
   renderShippingMethods,
+  renderShippingStatus,
   renderTermsAndConditions,
   unmountEmptyCart,
 } from './containers.js';
@@ -69,10 +80,11 @@ import {
   TERMS_AND_CONDITIONS_FORM_NAME,
 } from './constants.js';
 
-import { rootLink } from '../../scripts/commerce.js';
-
-// Success block entry point
-import { renderOrderSuccess } from '../commerce-checkout-success/commerce-checkout-success.js';
+import {
+  fetchPlaceholders,
+  rootLink,
+  SUPPORT_PATH,
+} from '../../scripts/commerce.js';
 
 // Initializers
 import '../../scripts/initializers/account.js';
@@ -238,12 +250,11 @@ export default async function decorate(block) {
   ]);
 
   async function displayEmptyCart() {
-    if (!emptyCart) {
-      emptyCart = await renderEmptyCart($emptyCart);
-      $content.classList.add(CHECKOUT_EMPTY_CLASS);
-    }
+    if (emptyCart) return;
 
-    removeOverlaySpinner(loaderRef, $loader);
+    emptyCart = await renderEmptyCart($emptyCart);
+
+    $content.classList.add(CHECKOUT_EMPTY_CLASS);
   }
 
   function removeEmptyCart() {
@@ -315,6 +326,56 @@ export default async function decorate(block) {
     }
   }
 
+  // Define the Layout for the Order Confirmation
+  async function displayOrderConfirmation(orderData) {
+    // Scroll to the top of the page
+    window.scrollTo(0, 0);
+
+    // Create order confirmation layout using fragments
+    const orderConfirmationFragment = createOrderConfirmationFragment();
+
+    // Create scoped selector for order confirmation fragment (following multi-step pattern)
+    const getOrderElement = createScopedSelector(orderConfirmationFragment);
+
+    // Get all order confirmation elements using centralized selectors
+    const $orderConfirmationHeader = getOrderElement(selectors.orderConfirmation.header);
+    const $orderStatus = getOrderElement(selectors.orderConfirmation.orderStatus);
+    const $shippingStatus = getOrderElement(selectors.orderConfirmation.shippingStatus);
+    const $customerDetails = getOrderElement(selectors.orderConfirmation.customerDetails);
+    const $orderCostSummary = getOrderElement(selectors.orderConfirmation.orderCostSummary);
+    const $orderGiftOptions = getOrderElement(selectors.orderConfirmation.giftOptions);
+    const $orderProductList = getOrderElement(selectors.orderConfirmation.orderProductList);
+    const $orderConfirmationFooter = getOrderElement(selectors.orderConfirmation.footer);
+
+    const labels = await fetchPlaceholders();
+    const langDefinitions = {
+      default: {
+        ...labels,
+      },
+    };
+    await initializers.mountImmediately(orderApi.initialize, { orderData, langDefinitions });
+
+    block.replaceChildren(orderConfirmationFragment);
+
+    await Promise.all([
+      renderOrderHeader($orderConfirmationHeader, { orderData }),
+      renderOrderStatus($orderStatus),
+      renderShippingStatus($shippingStatus),
+      renderCustomerDetails($customerDetails),
+      renderOrderCostSummary($orderCostSummary),
+      renderOrderProductList($orderProductList),
+      renderOrderGiftOptions($orderGiftOptions),
+    ]);
+
+    // Create footer content using fragments
+    $orderConfirmationFooter.innerHTML = createOrderConfirmationFooter(rootLink(SUPPORT_PATH));
+
+    const $continueButton = selectors.orderConfirmation.continueButton;
+    const $orderConfirmationFooterBtn = $orderConfirmationFooter.querySelector($continueButton);
+
+    await renderOrderConfirmationFooterButton($orderConfirmationFooterBtn);
+  }
+
   async function handleCheckoutInitialized(data) {
     if (isEmptyCart(data)) {
       await displayEmptyCart();
@@ -360,7 +421,7 @@ export default async function decorate(block) {
 
     window.history.pushState({}, '', url);
 
-    await renderOrderSuccess(block, { orderData });
+    await displayOrderConfirmation(orderData);
   }
 
   events.on('authenticated', handleAuthenticated);
